+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+-   Added `parseRawCellToConfigDict`, `cellToConfigDict` and `configDictToCell` utility functions to convert between a raw cell and a config dictionary
+
 ## [0.0.1] - 2025-04-02
 
 ### Changed